+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+- Removed join function
+- Add an interface to create your own text components
+- Add a custom decoration to text components
+- Fixed some bugs
+
 ## 0.0.1
 
 - TODO: Publish package.